Updating the CONTRIBUTING landing page
A new set of instructions for how to contribute to the storlet project, a lot
of the information is covered in the infra dev wiki but its more consolidated
in this form and potentially easier to pick up.  I will add more interesting
rebasing scenarios in the future as that's usually one of the trickier steps for
newer folks.

Change-Id: I319febd17eba48844703ccacaec3e67c5cc2b6a6