Refactor Storlet Handler middleware
For now, StorletHandlerMiddleware has too many if branches
for handling the request. (e.g. where to execute storlet, when
to execute, so on) However it decrease the maintenacibility
because it's not obvious where the statement belongs.

In this patch, split the handling sequence into mainly 2 Handler
classes called StorletProxyHandler and StorletObjectHandler which
inherit BaseStorletHandler.

BaseStorletHandler consists of common component for both chiled classes.
(e.g. is_storlet_exection) And then chiled classes cover for their own
requet handling.

Note that you may be able to find some odd code around header handling
but basically this code keep *original* handling not to modify current
behavior so if you find unnecessary code with this, please contribute
to improve the code in future work :-)

Co-Authored-By: Takashi Kajinami <kajinamit@nttdata.co.jp>

Change-Id: I1de30b631c9ad6321ee837bbf9e6ae96eb78d181