Extract ver/acc/con/obj instead of self._get_vaco
While refactoring in previous work, we find a lot of gateway
method (and logging message) need a part of request path like
account, container, object. In the work, we employed self._get_vaco
method which parses the path and returns a tuple
(version, account, container, object) to get the part of path with holding
the older behavior for clarifying what was refactored.

And now, we can focus on to replace whole self._get_vaco into self.vars.

Basically 2 changes are in this patch:

1. Merge every self._get_vaco into BaseStorletHandler:
self._get_vaco will be called when the handler is init and extract into
each variables to be able to be referenced directly.

2. Tied up request information to Handler
The variables like self.account will be duplicated information with
actual request instance and they can be maintained as different
variables. (i.e. if set account name directly like "self.account='acc'",
it might be different the account in self.request.path)

To prevent such an unexpected inconsistency, this patch add the wrap of
"self.request" variables to extract the ver/acc/cont/obj value each time
when the request variable assigned. And also this disallowed the direct
assinment for ver/acc/cont/obj in the Handler itself.

*bonus*
- Fix to raise NotImplementedError instead of NotImplemented (which is
  just a Type!)

Change-Id: Iccf0487d6aafd462f1290595d645966d0059474e