Move verify_access from storlet_gateway to storlet_middleware
This patch moves verify_access, which is a function specific,
to swift from gateway module to middleware module.

NOTE:
This patch also changes header keys, which is included in the
request between proxy and object, from X-Object-Meta-Storlet-Key
to X-Storlet-Key, to make it clear that they are just used to
Storlet execution.

Change-Id: I887b44176549bd7f0772864d3a4f3b65b91e4ad1