Working with a JSON cluster configuration
Various components in Storlets require to know the cluster config.
Examples are:
1. The functional tests need to know the keystone auth uri
2. The installation needs to know where are the proxies / object servers
3. The management scripts need to know about all the nodes.

Now all information is expected to be in a single json file used by all components
There is a template of such file in install/storlets/templates/cluster_config which
is used for the s2aio installation.

changes include:
1. Adapting the ansible scripts to use a dynamic inventory that uses the json
2. refactoring setup_functional_test.sh moving work to
   /install/storlets/install_storlets.sh
3. general cleanups and fixes to  the ansible scripts

Change-Id: I2810f01a977c40870cada476a9c54ce4a7bdcc1a