Refactor Storlet protocol classes
Previously, we have multiple Storlet protocol classes, corresponding
to each execution place, each methods.
This patch refactors these classes, and introduces a common protocol
class, which is used for all storlet executuions.
This patch also makes sure that we use a iterator as stream in
Storlet Request classes, as which we previously use file descriptor
only for GET.

Change-Id: If763da213e02829d492dc639037323f548224c68