Adding more information to storlet_handler info
In order to know if a certain functionality is
supported throught storlets, it is not enough
to know if the storlet_handler exists in the
pipeline (which is what currently covered by info)
but we also need the ability to know which storlets
are installed. Thus, we want to export the name
of the container holding the storlets.

It might be that gateways other then the current one
would store the storlets in a different manner, thus
we also export the class name of the gateway implementation

A usage of this is the spark use case, where we test
if the relevant storlet exists in order to decide whether
to activate the pushdown or not

This involves adding a new test and somewhat refactoring
the config parser.

Bonus: Fixing the ant deploy_host_engine and deploy_container_engine
that got broke with recent installation refactoring

Change-Id: I971244ef048ea4f580e95ac5be92b3768fd60929