Refactoring SBusDatagram
The SBus datagram is a serialization and de-serialization object that
is written both on python and in java. It serializes commands that go
from the middleware on the host side or the factory (client) to the
docker container (server). As such it has two logical components:
ClientSBusOutDatagram and ServerSBudInDatagram. In the past we also had
commands going in the opposite direction which required:
ServerSBusOutDatagram and ClientSBusInDatagram

The current implementation had all logical parts in the same object
which made it very difficult to maintain. some fields were set and other not
depending on the usage (in cient or in server), and the accessors were a mess.

This patch introduces independent classes for:
ClientSBusOutDatagram - used by the middleware to send commands to the
daemon_factory and the storlet daemons. We give python implementation.
ServerSBudInDatagram - used by the daemon_factory and storlet daemons
to parse the messages creasted by the ClientSBusOutDatagram. As such
we have both a python and java implementations.

While doing the refactoring style and liscense headers were cleaned up.
Also, all serialized data (except fds) are strings, which make various
conversions done before to be redundant. Overall there is less code
and is much simpler.

Bonus 1: Initial work to be able to see logs of the storlet daemons.
this was missing for a long time now. Currently one has to do:
docker exec -it <container id> /bash/bin and see /var/log/SDaemon.log
this needs to be further improved. Ideally we could take a file descriptor
given by the host and build a log4j on top of it. Partr of the work
was to move the logback.xml installation from the generic java image to the
storlet image as it is storlets related.

Bonus 2: fixes to the deploy_container_engine role. There are still problems
and for the script to work the previous container must be deleted.

Change-Id: Ide59797f451e238b1c424f5ee811d5f1eca616b7