Splitting fd metadata to system and user
Currently, the fd metadata being passed to the storlet side
is a single flat dictionary, that consists of both storlet internal
system md and user md, which is swift object md.

This structure is not extendable, esp. with the additional
system metadata we need for ranges. This patch splits the
metadata into two sub dictionaries:
1. "storlets"
2. "storage"
In the future as we also pass swift sysmeta we should
further split the "storage" part to system and user.
Note, however, that this may require update to existing
storlets or introduce a new API version.

Bonus: ability to run functest in a dev / jenkins flavors
       where dev flavor skips slow tests e.g. 1GB put
       which also fills up the dev env device...
       docs are updated accordingly.

Change-Id: I88ecba46a31fa8b201ff3b50d1449aae32c7ef82