Single range request on obj.
In the SQL pushdown scenario from Hadoop/Spark all requests
are single range requests. Currently, all range requests
are executed on the proxy by translating the X-Storlet-Range
request to Range request, letting the GET return the required Range
and process that range on the proxy.

To better facilitate the SQL push down scenario, we want to run
the storlet on the object node and do the filtering really at the source
Recall, that on the object node we pass the actual object file fd
to the storlet. We introduce a new input stream at the storlet side
that exposes to the storlet only the required range from that fd.

Thus without any change to the storlet API, in case the original
request has a range, the storlet will 'see' only that range

Bonuses: fixing the disastorous tab/spaces mixed formatting
         of sbus.c, fixing logback.xml to log in the container
         where thete are permissions

Change-Id: I25eeaa38c886608cdbebcf65b4a95b3f1cd270f8