Add unittests and cleanup codes
This patch works around StorletGatewayDocker.invocation_flow and
RnTimePaths.

To assert *REAL* behavior, this patch adds a few unittests to run in
real os dir. This patch also add an unit test to assert the
communication of docker container (basically wrap up the response from
the container with mock) and check the real handling.

This could be the first step to emurate real environment in *UNIT TEST*
rather than functional tests for easy reproducing and finding real bugs.

In the future work, we could add a temporary server process communicating
with DockerGateway instance via fds.

Change-Id: I5fbd7cd415053b7913310973f71ad5a7ab48faa0