Enable 2 or more input for a storlet app
Some application may want to get multiple objects for storlet exectuion.

For example, we'd like to get the difference from text-A and text-B,
both exist in the storage side. Currently, storlet can not reach out the
use case so that we have to retrieve the objects to local by hand, and
then should check the difference. However, imagine storlet can take
multiple input, what we should do is just a making an application to
return the diff from 2 input stream.

What's new:

This patch adds a new API header "X-Storlet-Extra-Resources" in the
request which enables multiple input streams to a storlet app.
X-Storlet-Extra-Resoureces allows the value formatted as a csv consists
of swift path ('/container/object') that is similar with swift copy
request. Note that currently only the path container + object is allowed.
(including account path is not yet supported) If you want to more
streams, you can add more paths to the csv. The object streams will be
appeared as other streams behind of the main stream. (main PUT or GET)
i.e. if you specified 2 extra streams in the header, 3 streams (1 main
and 2 extras) will be supplied for your apps.

Example is like:
GET /con/main
X-Storlet-Extra-Resources: /con/oA, /con2/oB

your app get StorletInputStream Array like,
[stream(/con/main), stream(/con/oA), stream(/con2/oB)]

How it works:

Storlet middleware will handle the X-Storlet-Extra-Resoureces header.
If it is comming from, storlet handler will try to get the object(s)
from backend storage and set the stream(s) into the storlet app via SBus
protocol. Each stream instance is just a input stream as well as main
object stream we are currently using for storlet apps.

Note that currently the extra resources is designed for running only
proxy-server. To run on the object-server, we need more complicated
works how to get the extra resources and this version *may* not have the
ristriction to prevent to run that app in an object-server
unfortunately.

TODO (it could be in follow up):
- Description docs for the new API
- More tests with/without X-Storlet-Run-On-Proxy

Change-Id: I3aa27d4df5bb151c09003a98209c7ccf91d25828