Fix COPY request
Currently COPY middleware is created in swift, and
"COPY" is allocated to the left of the "storlets_handler"
in the pipeline of proxy-server.
Therefore COPY request is processed in the COPY middleware.

In this case , there is the following problem.
1. Failed content_length check
 Because storlets doesn't set content_length.
2. COPY request handles storlets twice
 storlets is processed in the GET request and the PUT request.

This patch moves storlet_handler to the left of
copy middleware, to process COPY request in storlet_handler,
as we did with previous swift.

Change-Id: I325e28451bf0ee1dead88e0cb6009d3f7265b677
Closes-Bug: #1606446