Creanup redundant README.md and README.rst
Once we have a couple of typs for RADME (.md and .rst) but it makes us
hard to maintain which is the newest updated and which one we should
refer as the right docs. This patch merges the two files into .rst which
is preferred by OpenStack projects.

Plus it fixes some link errors in the doc but probaly we still need to
update this one as the correct docs in follow-up.
(I was trying to do that but it was hard finding if which ones are current
correct docs or not)

Change-Id: I31b24159d8c2200af91980357ee7deeb13f95707