Set os.environ parameters in subprocess
Previously, we set os.environ parameters in the main process of
daemon_factory, but it is not proper because that setting can affect
whole of the base process.

This patch makes daemon_factory to set os.environ value when starting
subprocess, to make sure nothing about the main process is changed.

Change-Id: I01465ba50ce37272ec3003d653d4c76a462cb3a6