New storlet for testing pushdown
PartitionsIdentityStorlet is a new storlet that has built in support
for cross record ranges: Spark tasks can break an object in the middle
of a record. This storlet mimics the mepreduce solution for that
allowing to bypass the mapreduce layer altogether when doing a pushdown
from Spark.

Bonus:
Two minor fixes in the install scripts

Change-Id: I8ac3b0512f7d5a4318e23e459e560c4b8f6835c2