Install Swift and the Storlet Engine on a Docker container.
This is much less intrusive than the current installation, since
most of modifications occur within the Docker container.
Another advantage is that, unlike the current installation, it
does not require passwordless-ssh to root account.

The behavior of Docker containers is consistent (unlike other
Ubuntu machines), so you could be fairly certain that the
installation would succeed.

The Docker container is built through a new option
to s2aio.sh called target that can take the values
of docker | host
Use docker to install everything in a Docker container
Use host to install on the host where the script is
executed.

Change-Id: Id693b2b16e668bffa525ae2aa0b4aa40f21067eb