Add validation logic for python storlets
This patch adds validation logic specific for python storlet files,
to support python as a language to write Storlet applications.

The validation logic is based on based on the following rule.
 1. the object should be a python file, whose name ends with .py
 2. X-Object-Meta-Storlet-Main should indicate the class name
    in the python file

Currently it only supports singile python file, but we also have
a plan to let it support bdist tar file in the future.

Implements: blueprint python-storlet
Change-Id: I58d0417d68f4da0afe4ef9dc5008a1c16a9b71a8