Pass storlet language from gateway to daemon factory
This patch enables docker gateway to properly pass the
value of 'X-Object-Meta-Storlet-Language' to daemon
factory, to let daemon factory start python storlet
daemon for python storlets.

NOTE:
This patch also changes the default log level in daemon
factory and sdaemon from TRACE, which is specific in java
logging, to DEBUG, which is common for Java and Python
logging.

Change-Id: I8be37bbdb7c3dfd4d11c855464ecf789932fa354