Use openjdk isntead of oracle jdk
We've used oracle jdk as a java runtime to run storlet java codes,
but it is not necesorry to use oracle jdk, and from some legal points
of view, it is better to use openjdk instead.

This patch makes changes to current scripts and files for installation,
to make them use openjdk.

NOTE:
Currently we add ppa:openjdk-r/ppa into apt repositoriy setting, but
we don't need that addition when we use Ubuntu 14.10 or later.

Closes-bug: #1618895
Change-Id: I4c65afd4edf6b0d8b5f5f83307a8129e67c2911d