Bump up ansible for linting
Previously we set a version cap about ansible only for linting,
but that caused the defference between ansible used in linting
and the other one used in actual deployment.

This patch bumps up ansible for lint testing, and make sure we
use latest version of ansible for linting and deployment.

Change-Id: I6475319c120fca1c6f44129c360336d1fb63fd5c