Get rid of default value setting in storlet_handler
Previously we set default value for some of config parameters
in storlet_handler, not in the place where we really use it.
In this patch, I set default values in the module which really
uses these values, to let us find easily what are the real
values set as defaults.

This patch also makes middleware configration and gateway
configration independent from the other, because we currently
do not have config parameters which is used by both of
middleware and gateway.

Change-Id: I05d865bd9da3ebe54629b241d2fccca44d69b8fa