Remove redundant passing container id in daemon factory
Currently we pass container id to DaemonFactory.main, but it
causes unnecessary passing container id to handler functions,
which never use the passed value.
The container id is never changed while the daemon factory
process is running, so we can define it when creating
DaemonFactory instance, and refer it as a class member.

Change-Id: Ica70ce474963aa53ac6379cce171fd31ec52db54