Pre-cutting a stable-newton branch
updating CHANGELOG with the release content
freezing the swift dependency version to 2.9.0
which is the Newton release. I will change that back to master
as soon as the branch is cut. I guess that we need to change this
procedure for future releases.

There seem to be no Newton Ubuntu repo for 14.04,
and so the Keystone dependency is left as Mitaka.
Yet another thing to resolve in future releases.

Change-Id: Icecde88b253a5f94c9fc42052b9b922feb459e0f