Using devstack for swift and keystone installation
This change gets rid of our proprietary scripts
for installing Swift and Keystone in favor of using
devstack.

So far we have used those scripts for:
1. The functional test gatejob
2. The s2aio.sh script

The merits are:
1. No need to maintain these ourselves.
   Specifically, this gives us Linux distros
   other than Ubuntu for free
2. Being more Openstack compliant. Specifically,
   avoid packages conflicts that arise from our
   script usage of apt and the storlets setup
   environemnt which uses pip.

The next step is to have a storlet devstack plugin
that would allow to setup a dev/test env for storlets
that is purely based on devstack.

It might be that once having a full devstack integration
we will clean the storlets repo from all ansible
scripts, and drop the goal for supplying a multi-cluster
installation support.

Change-Id: I4731fe7369bc86bc626f08833aba08e5ff1d7184