Add start / stop to s2aio
s2aio uses a loop back device over a local file
as Swift device. This device may get corrupted
when rebooting the host while swift is running.
Adding the start and stop helps mitigate the problem

Also, with the recent move to devstack as the
installation script for Swift and Keystone
starting the swift and keystone services
have changed. Again, the s2aio start helps
here.

Change-Id: I3e3884a1a04888be577fe4f92d471a65f6f5ce5b