Add functional tests for python storlet execution with parameters
This patch adds some funcitonal tests to test capability of storlet
engine to run python written applications, especially to test that
it can handle given parameters in request queries or headers.

It introduces the same test cases as the following ones for java
written storlets, so that we can have the same test coverage for
python written storlets.
<tests/functional/java/test_identity_storlet.py>

Change-Id: I70d36767be7df7d3be0f5b4339939a1d524c3f11