Extension for IPython/Jupyter notebook
This patch adds ipython extension that can be useful to implement/deploy
the testing storlet apps with the actual storlet/swift cluster.

Quick Start:

1. install storlet package to your development environment (e.g. python
   setup.py install)
2. set your storlet environ to your environment values (e.g. OS_AUTH_URL)
   that may be in your .bashrc if you are using s2aio.sh for your instance.
3. Use ipython or jupyter notebook and set following load module magic.
   %load_ext storlet.tools.extensions.ipython
   (or %reload_ext storlet.tools.extensions.ipython)
4. Use %%stolretapp magic for your cell

For more detail, please see IPython reference,
https://ipython.org/documentation.html

Remaining Tasks (expecting as future works):

`--dry-run` using sort of https://review.openstack.org/#/c/409003/ will be
supported in the future. That is not in the scope of this patch.

auto_generated docs for the extension said that the extension
is used as '%storletapp' but it's wrong. The cell command should be
'%%storletapp' (doubled %).

Change-Id: Ie7e046f2cc0500f4c05e7e96e05e3a1a42ac3dbb