Use json for response sent through service out fd
Now we use very simple format using ':' as a separator return
task status and a message related to the task operation, but
that protocol makes it difficult to add more informations into
the response.

This patch changes its format to json, so that we can include
more information keys/values into the response easily.

Change-Id: If2d609fb2ba2187e121e9cadd281017af46a6661