Getting all storlet related code out of the container
One of the biggest hurdles in the storlets devops code
is the need to update the container image each time
the storlet related code changes.

A previous patch [1] already made sure that
all the container necessary code is installed on the host
machine in:
/usr/local/lib/storlets
/usr/local/lib/python2.7/dist-packages/storlets

This patch completes the work by:
1. Treat the scripts daemon, daemon-factory, init_container.sh
   accordingly: placing them in /usr/local/libexec/storlets
2. Mount the above-mentioned dirs in ro mode
3. Cleanup the ansible code that installs everything
   in the container image
4. Update the installation docs accordingly

[1] https://review.openstack.org/#/c/370332/

Change-Id: I7da46c1554df2bbbff2ccff12d82d8fbe158aa22

Getting all storlet related code out of the container

One of the biggest hurdles in the storlets devops code
is the need to update the container image each time
the storlet related code changes.

A previous patch [1] already made sure that
all the container necessary code is installed on the host
machine in:
/usr/local/lib/storlets
/usr/local/lib/python2.7/dist-packages/storlets

This patch completes the work by:
1. Treat the scripts daemon, daemon-factory, init_container.sh
   accordingly: placing them in /opt/storlets/bin
2. Mount the above-mentioned dirs in ro mode
3. Cleanup the ansible code that installs everything
   in the container image
4. Update the installation docs accordingly

[1] https://review.openstack.org/#/c/370332/

Change-Id: I7da46c1554df2bbbff2ccff12d82d8fbe158aa22