PY3: ensure bytes format in sbus conversation
Originally, that was just str, that can be compatible with
char in python c_types. However, in python3, str is not a
byte format but bytes. This patch is intended to encupsulate
any data format (path name, metadata) into SBus module that
tranforms them from str to bytes at the send time. (receive
will work as well from bytes to str) Thus, any other storlet
modules are able to use them as str as well as ever.

Change-Id: I542db555798e7ef67bf0db9d98d1c14578118f86