Avoid broken http header including white space
The execution result header in user metadata looks invalid
and it cuases the hang in the func CI tessting.
So let's remove the dangerous white space in the http
header key with this patch.

The skip test removed in this patch has been since
https://review.opendev.org/#/c/693549 . The skipping
purpose was to progress py3 migration (we'd like to find
any other issues than this at the time) so the test is still
worth, then this fix enables to pass the test.

Change-Id: Ib41735d8801f0c99ed0563213dcf414fefb5cbb2