Add warning log when restarting docker container
When a storlet application or its dependencies are updated, we first
try restarting storlet daemon, but if it fails then we restart the
whole container, which might cause ungraceful shutdown of the other
storlet daemon in the same container.

This change adds a warning log so that operators can understand how
many times such stop failure have happened in the deployment.

Change-Id: I2843bf1118505a2e32e1f80e7329ed73de1cd62d