Run agent processes by the swift user
Currently agent processes inside storlet containers are started by
the root user and inpersonate the swift user. However this is not
secure and might allow unexpected usage of the root user.

This change ensures the agent processes are started by the swift user
so that it gets limited access from beginning.

Change-Id: Iea1c6219d2a4f889623a3a0f36698cdb67663f55