Remove iPython extension
This has been unmaintained and its functional tests have been broken
for some time. Because the current upper version is not compatible with
Python 3.12 and we aren't sure if we can bump the cap safely, let's
remove the extension.

Also replace the openstack CLI by the swift CLI to resolve the current
CI failure.

Change-Id: I1a2f908fef4b660686d75ea6e8330287b423cd8b